Refatora function.py para tools.py
